Use perf archive and stop generating per-CPU reports
The generation of per-CPU perf reports can take a long time to generate
when there are many CPUs present, since it is performed serially to avoid
memory usage bloat processing them all in parallel.

We have moved all of the logic for post-processing perf reports to the
perf-postprocess Perl script, and now use perf archive to capture all
the necessary data to process the perf data off host.